refactor: move terminal visuals into appearance settings
